+++ Pipfile
@@ -5,6 +5,7 @@ name = "pypi"
 
 [packages]
 flask = "*"
+pandas = "*"
 "psycopg2-binary" = "*"
 
 [dev-packages]
