+++ scripts/service_provider_data.csv
@@ -1,0 +1,3 @@
+Service Provider,Website,Report Year,PDF link,Operating Budget (Expenses),Clients Served,Staff Count,Main Address,Satellite Addresses,Notes,Data Collection Notes,Questions
+Youthlink,www.youthlink.ca,2017,https://youthlink.ca/wp-content/uploads/2017/12/Good-News-Everyone-ELECTRONIC.pdf,5692028,8711,70,"636 Kennedy Road, Scarborough, Ontario M1K 2B3",,,,
+Youth Services Bureau of Ottawa,www.ysb.ca,2016,https://www.ysb.ca/wp-content/uploads/2017/07/final-ar-2016-english.pdf,26834132,1000,350,"2675 Queensview Drive, Ottawa, Ontario K2B 8K2",,,,
